chore: move around the legacy parts so that they can be deleted
