+++ plugins/scaffolder/src/legacy/TemplateEditorPage/TemplateFormPreviewer.tsx
@@ -34,8 +34,8 @@ import useAsync from 'react-use/lib/useAsync';
 import yaml from 'yaml';
 import { type LayoutOptions } from '@backstage/plugin-scaffolder-react';
 import { TemplateEditorForm } from './TemplateEditorForm';
-import { TemplateEditorTextArea } from './TemplateEditorTextArea';
 import { LegacyFieldExtensionOptions } from '@backstage/plugin-scaffolder-react/alpha';
+import { TemplateEditorTextArea } from '../../components/TemplateEditorPage/TemplateEditorTextArea';
 
 const EXAMPLE_TEMPLATE_PARAMS_YAML = `# Edit the template parameters below to see how they will render in the scaffolder form UI
 parameters:
